requirements.txt
asgiref==3.4.1
Django==3.2.12
Pillow==7.0.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
Django==3.2.7